+++ src/endfield_factory_compiler/__init__.py
@@ -2,6 +2,7 @@
 
 from .compiler import CompilationResult, compile_project
 from .execution import ExecutionOptions
+from .model import FloorplanSearchOptions, FloorplanSearchResult
 from .pack import load_project, load_region_pack
 from .routing import GridAStarRouter
 from .routing_backend import RouterBackend, RoutingStats
@@ -9,6 +10,8 @@ from .routing_backend import RouterBackend, RoutingStats
 __all__ = [
     "CompilationResult",
     "ExecutionOptions",
+    "FloorplanSearchOptions",
+    "FloorplanSearchResult",
     "GridAStarRouter",
     "RouterBackend",
     "RoutingStats",
@@ -17,4 +20,4 @@ __all__ = [
     "load_region_pack",
 ]
 
-__version__ = "0.4.2"
+__version__ = "0.4.3"